T4-01: ReplanQueue 이진 힙
라운드로빈 FIFO 스텁을 점수순 최대 힙으로 교체한다. 고정 크기 · 할당 0 ·
_heapPos 로 중복 검출 후 점수 갱신 · 포화 시 최하위 축출.

점수 타입이 int → float 로 넓어져 호출부 3곳을 같이 고쳤다
(CognitionScheduler.Score · InterruptMatcher · PlanExecutor.ReplanOnFailScore).
인터럽트는 TryEnqueueUrgent(1000 + urgency) 를 쓴다 — rule.Urgency 를 그대로
넣으면 최댓값 100 이라 일반 재계획과 뒤섞인다.

docs/14 §2 초안 세 곳을 같은 커밋에서 정정했다 (사유는 TASKS.md §3).

+++ src/Npc.Runtime/CognitionScheduler.cs
@@ -113,12 +113,12 @@ public sealed class CognitionScheduler
     }
 
     /// <summary>
-    /// 재계획 우선순위 점수. P1 은 LOD 만 본다 (가까울수록 급하다).
-    /// 정식 가중치(플레이어 근접 · 플랜 노후 · 전제 이탈 · 긴급도)는 T4-01 이 넣는다.
+    /// 재계획 우선순위 점수. 밴드만 본다 (가까울수록 급하다).
+    /// 정식 가중치(플레이어 근접 · 플랜 노후 · 전제 이탈 · 긴급도)는 T4-02 의 <c>ReplanScorer</c> 다.
     /// </summary>
-    private static int Score(int npc, int band)
+    private static float Score(int npc, int band)
     {
         _ = npc;
-        return (LodBandSet.BandCount - band) * 10;
+        return (LodBandSet.BandCount - band) * 10f;
     }
 }
